feat: use subdomain gateway for preview url
ipfs cid-in-subdomain gateways are preferred now as they preserve the origin isolation of each site by using the cid as part of the domain. Each root is isloated from every other root.

Also it means your site is mounted on / rather than /ipfs so root-relative urls will still work, which is a big deal.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>