feat: use subdomain gateway for preview url (#3)
ipfs cid-in-subdomain gateways are preferred now as they preserve the origin isolation of each site by using the cid as part of the domain. Each root is isloated from every other root. It means your site is mounted on / rather than /ipfs so root-relative urls will still work too, which is a big deal.

Also adds the gateway url as an action output, for nice nice.